+++ heylogs-api/src/main/java/internal/heylogs/GitHostLink.java
@@ -1,0 +1,14 @@
+package internal.heylogs;
+
+import lombok.NonNull;
+
+public interface GitHostLink {
+
+    @NonNull String getProtocol();
+
+    @NonNull String getHost();
+
+    int getPort();
+
+    int NO_PORT = -1;
+}