+++ heylogs-api/src/main/java/internal/heylogs/GitHostRef.java
@@ -1,0 +1,8 @@
+package internal.heylogs;
+
+import lombok.NonNull;
+
+public interface GitHostRef<T extends GitHostLink> {
+
+    boolean isCompatibleWith(@NonNull T link);
+}