+++ heylogs-api/src/main/java/internal/heylogs/github/GitHubRules.java
@@ -2,13 +2,13 @@ package internal.heylogs.github;
 
 import com.vladsch.flexmark.ast.Link;
 import com.vladsch.flexmark.util.ast.Node;
+import internal.heylogs.GitHostSupport;
 import lombok.NonNull;
 import nbbrd.design.MightBeGenerated;
 import nbbrd.design.VisibleForTesting;
 import nbbrd.heylogs.Failure;
 import nbbrd.heylogs.spi.Rule;
 import nbbrd.heylogs.spi.RuleBatch;
-import nbbrd.io.text.Parser;
 import nbbrd.service.ServiceProvider;
 
 import java.util.stream.Stream;
@@ -28,6 +28,12 @@ public enum GitHubRules implements Rule {
         public Failure validate(@NonNull Node node) {
             return node instanceof Link ? validateGitHubPullRequestRef((Link) node) : NO_PROBLEM;
         }
+    },
+    GITHUB_MENTION_REF {
+        @Override
+        public Failure validate(@NonNull Node node) {
+            return node instanceof Link ? validateGitHubMentionRef((Link) node) : NO_PROBLEM;
+        }
     };
 
     @Override
@@ -42,46 +48,59 @@ public enum GitHubRules implements Rule {
 
     @VisibleForTesting
     static Failure validateGitHubIssueRef(Link link) {
-        GitHubIssueLink expected = Parser.of(GitHubIssueLink::parse).parse(link.getUrl());
-        if (isGitHubLink(expected) && expected.getType().equals(GitHubIssueLink.ISSUES_TYPE)) {
-            GitHubIssueRef found = Parser.of(GitHubIssueRef::parse).parse(link.getText());
-            if (isCompatibleRef(found, expected)) {
-                return Failure
-                        .builder()
-                        .rule(GITHUB_ISSUE_REF)
-                        .message("Expecting GitHub issue ref " + expected.getIssueNumber() + ", found " + found.getIssueNumber())
-                        .location(link)
-                        .build();
-            }
-        }
-        return NO_PROBLEM;
+        return GitHostSupport.validateRef(
+                GitHubIssueLink::parse,
+                GitHubIssueRef::parse,
+                GitHubRules::isExpectedIssue,
+                link, GITHUB_ISSUE_REF,
+                GitHubRules::getIssueMessage);
+    }
+
+    private static boolean isExpectedIssue(GitHubIssueLink expected) {
+        return expected.getType().equals(GitHubIssueLink.ISSUES_TYPE) && expected.getHost().equals("github.com");
+    }
+
+    private static String getIssueMessage(GitHubIssueLink expected, GitHubIssueRef found) {
+        return "Expecting GitHub issue ref " + (found.isShort() ? GitHubIssueRef.shortOf(expected) : GitHubIssueRef.fullOf(expected)) + ", found " + found;
     }
 
     @VisibleForTesting
     static Failure validateGitHubPullRequestRef(Link link) {
-        GitHubIssueLink expected = Parser.of(GitHubIssueLink::parse).parse(link.getUrl());
-        if (isGitHubLink(expected) && expected.getType().equals(GitHubIssueLink.PULL_REQUEST_TYPE)) {
-            GitHubIssueRef found = Parser.of(GitHubIssueRef::parse).parse(link.getText());
-            if (isCompatibleRef(found, expected)) {
-                return Failure
-                        .builder()
-                        .rule(GITHUB_PULL_REQUEST_REF)
-                        .message("Expecting GitHub pull request ref " + expected.getIssueNumber() + ", found " + found.getIssueNumber())
-                        .location(link)
-                        .build();
-            }
-        }
-        return NO_PROBLEM;
+        return GitHostSupport.validateRef(
+                GitHubIssueLink::parse,
+                GitHubIssueRef::parse,
+                GitHubRules::isExpectedPullRequest,
+                link, GITHUB_PULL_REQUEST_REF,
+                GitHubRules::getPullRequestMessage);
     }
 
-    private static boolean isGitHubLink(GitHubIssueLink expected) {
-        return expected != null && expected.getHost().equals("github.com");
+    private static boolean isExpectedPullRequest(GitHubIssueLink expected) {
+        return expected.getType().equals(GitHubIssueLink.PULL_REQUEST_TYPE) && expected.getHost().equals("github.com");
     }
 
-    private static boolean isCompatibleRef(GitHubIssueRef found, GitHubIssueLink expected) {
-        return found != null && !found.isCompatibleWith(expected);
+    private static String getPullRequestMessage(GitHubIssueLink expected, GitHubIssueRef found) {
+        return "Expecting GitHub pull request ref " + (found.isShort() ? GitHubIssueRef.shortOf(expected) : GitHubIssueRef.fullOf(expected)) + ", found " + found;
     }
 
+    @VisibleForTesting
+    static Failure validateGitHubMentionRef(Link link) {
+        return GitHostSupport.validateRef(
+                GitHubMentionLink::parse,
+                GitHubMentionRef::parse,
+                GitHubRules::isExpectedMention,
+                link, GITHUB_MENTION_REF,
+                GitHubRules::getMentionMessage);
+    }
+
+    private static boolean isExpectedMention(GitHubMentionLink expected) {
+        return expected.getHost().equals("github.com");
+    }
+
+    private static String getMentionMessage(GitHubMentionLink expected, GitHubMentionRef found) {
+        return "Expecting GitHub mention ref " + GitHubMentionRef.of(expected) + ", found " + found;
+    }
+
+    @SuppressWarnings("unused")
     @MightBeGenerated
     @ServiceProvider
     public static final class Batch implements RuleBatch {